Move tests for loading location items from LocationDetailView.spec.tsx to Table.spec.tsx
